Move `consts` package out of `iptables` directory
The constants in the consts package are no longer specific to iptables,
so moving the package to a higher-level directory improves organization
and better reflects its broader usage across the codebase
