+++ pkg/transparentproxy/config/config_executables.go
@@ -17,7 +17,7 @@ import (
 	"golang.org/x/exp/maps"
 	k8s_version "k8s.io/apimachinery/pkg/util/version"
 
-	. "github.com/kumahq/kuma/pkg/transparentproxy/iptables/consts"
+	. "github.com/kumahq/kuma/pkg/transparentproxy/consts"
 	util_maps "github.com/kumahq/kuma/pkg/util/maps"
 )
