+++ pkg/transparentproxy/iptables/parameters/protocol.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 
 	"github.com/kumahq/kuma/pkg/transparentproxy/config"
-	"github.com/kumahq/kuma/pkg/transparentproxy/iptables/consts"
+	"github.com/kumahq/kuma/pkg/transparentproxy/consts"
 )
 
 var _ ParameterBuilder = &ProtocolParameter{}
