+++ pkg/transparentproxy/iptables/setup.go
@@ -9,8 +9,8 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/kumahq/kuma/pkg/transparentproxy/config"
+	"github.com/kumahq/kuma/pkg/transparentproxy/consts"
 	"github.com/kumahq/kuma/pkg/transparentproxy/iptables/builder"
-	"github.com/kumahq/kuma/pkg/transparentproxy/iptables/consts"
 )
 
 func Setup(ctx context.Context, cfg config.InitializedConfig) (string, error) {
